Struct custodian_password::ServerLogin

pub struct ServerLogin(_);

Starts the login process on the server.

Implementations

impl ServerLogin

pub const fn config(&self) -> Config

Returns the Config associated with this ServerRegistration.

pub fn login(
config: &ServerConfig,
file: Option<ServerFile>,
request: LoginRequest
) -> Result<(Self, LoginResponse)>

Starts the login process. The returned LoginResponse has to be send back to the client to drive the login process. See ClientLogin::finish().

If a client is registered, pass the appropriate ServerFile, pass None otherwise. Passing None simulates a login attempt in a way that doesn’t let an attacker determine if a corresponding client is registered or not.

Errors

Error::Config if ServerConfig, ServerFile or LoginRequest were not created with the same Config

Error::ServerConfig if ServerFile was not created with the same ServerConfig

Error::Opaque on internal OPAQUE error

pub fn finish(self, finalization: LoginFinalization) -> Result<()>

Finishes the login process.

Errors

Error::Opaque on internal OPAQUE error.